Added changes to fix data reorder issue for non-powers-of-two arrays
Issue was related to the data padding implemented by accell in katsdpsigproc.
For certain array sizes padding would be added but when doing the pointer arithmentic
in the GPU this extra padding was not factored in. A simple fix is to specify that accel
should only use exact amount of memory as required. This is achieved through
e.g. accel.Dimensions(self.n_ants, exact=True)(do for all dimensions) for input and
output shapes. A few smaller changes also added (specifying input shape for CPU reorder
and not basing it on the template due to using the Dimension object) as well as some numpy
tweaks for type converstion in unit test. Basic code cleanup tookplace as well.