Scratchpad version of creating a multiply operational sequence for GPU execution.
Created a framework to add two multiply operations into an operational sequence. This
is based on the fill-reduce and multiply examples in katsdpsigproc docs. It differs in that
the GPU source is custom rather than using an existing katsdpsigproc operation (fill and hreduce).
This forms the base of the beamform as the first operation is a reorder then a custom tensorcore
complex multiply.