Add further updates to SKARAB DSim
In getting the simplest of tests to work, I have refined the process in
which the SKARAB DSim object is created and configured. The SKARAB DSim
is now able to 'program', start and stop data transmission.

I also removed the logging from utils.py as the basicConfig was creating
duplicate logs in the terminal.

Contributes to: NGC-394.