Add programming infrastructure for SKARAB DSim
It is, admittedly, very unpolished. But the main structure of the
process is in there now and has been tested successfully. As always,
places where code has been left in commented out is largely because it
requires a bit more digging in to why it doesn't work as expected.

The inclusion of progska needs to be cleaned up, it was just pasted
as-is from casperfpga.

Contributes to: NGC-394.