Make the entire system more robust
It's now almost an end-to-end kind of thing, writing the tests will get
you a PDF generated (presuming all the requirements are installed).

I need to work on the templating for the frontmatter, and getting
timestamps and other various things in place. But I'm happy with the
progress so far.